zavrsen projekat
izmenjen css,html,dodate neke funkcije,ubacen regEx.

+++ assets/js/contact.js
@@ -4,6 +4,7 @@ function setCookie(name, value, exDays) {
     today.setTime(today.getTime() + 1000 * 60 * 60 * 24 * exDays);
     document.cookie = name + "=" + value + "; " + "expires=" + today.toUTCString();
   }
+  
   /*brisanje kolacica */
 function deleteAll() {
   setCookie("cart", null, -1);
@@ -11,8 +12,19 @@ function deleteAll() {
 
 buyButton = document.getElementById("buyButt");
 buyButton.addEventListener('click', () => {
+  let formIme = document.getElementById("form_name1").value;
+  let formPrezime = document.getElementById("form_lastname1").value;
+  let formEmail = document.getElementById("form_email1").value;
+  let formAdress = document.getElementById("form_adress").value;
+  let formCity = document.getElementById("form_city1").value;
+  let formMessage = document.getElementById("form_message").value;
+
+  if(formIme.length==0|| formPrezime.length==0 || formEmail.length==0 || formMessage.length==0 || formCity.length==0 || formAdress.length==0){  
+    window.alert("Neispravan unos");
+    return;
+  };
     deleteAll();
-    window.alert("Uspesna kupovina");
+    window.alert("Uspešna kupovina");
     window.location = '/projects.html';
 });
 
@@ -22,6 +34,7 @@ logoutbutton.addEventListener('click', () => {
   setCookie("login", null, -1);
   window.location = '/projects.html';
 });
+
 /*rasclanjavanje kolacica za logovanje  */
 const parsirajLogin = () => {
   const cookies = document.cookie
@@ -32,7 +45,8 @@ const parsirajLogin = () => {
   if (cookies) {
     loginInformacije = JSON.parse(cookies.split("=")[1]);
   }
-/*uradjeno da se vidi ko je logova i da povuce informacije o korisniku */
+
+  /*uradjeno da se vidi ko je logova i da povuce informacije o korisniku */
   if (loginInformacije) {
     document.getElementById('logOut').hidden = false;
     console.log(JSON.stringify(loginInformacije));
